Fix build and tests
