+++ extensions/kube-object-event-status/package.json
@@ -1,6 +1,6 @@
 {
   "name": "kube-object-event-status",
-  "version": "0.0.1",
+  "version": "5.3.0-latest.1643044317859.1643044464147",
   "description": "Adds kube object status from events",
   "renderer": "dist/renderer.js",
   "lens": {
