+++ extensions/node-menu/package.json
@@ -1,6 +1,6 @@
 {
   "name": "lens-node-menu",
-  "version": "0.0.1",
+  "version": "5.3.0-latest.1643044317859.1643044464147",
   "description": "Lens node menu",
   "renderer": "dist/renderer.js",
   "lens": {
