+++ src/main/__test__/cluster.test.ts
@@ -51,8 +51,7 @@ describe("create clusters", () => {
 
     const di = getDiForUnitTesting({ doGeneralOverrides: true });
 
-
-    const mockOpts = {
+    mockFs({
       "minikube-config.yml": JSON.stringify({
         apiVersion: "v1",
         clusters: [{
@@ -74,9 +73,7 @@ describe("create clusters", () => {
         kind: "Config",
         preferences: {},
       }),
-    };
-
-    mockFs(mockOpts);
+    });
 
     await di.runSetups();
