--- src/renderer/components/+extensions/install-from-select-file-dialog/install-from-select-file-dialog.injectable.ts
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) OpenLens Authors. All rights reserved.
- * Licensed under MIT License. See LICENSE in root directory for more information.
- */
-import { getInjectable, lifecycleEnum } from "@ogre-tools/injectable";
-import { installFromSelectFileDialog } from "./install-from-select-file-dialog";
-import attemptInstallsInjectable from "../attempt-installs/attempt-installs.injectable";
-import directoryForDownloadsInjectable from "../../../../common/app-paths/directory-for-downloads/directory-for-downloads.injectable";
-
-const installFromSelectFileDialogInjectable = getInjectable({
-  instantiate: (di) =>
-    installFromSelectFileDialog({
-      attemptInstalls: di.inject(attemptInstallsInjectable),
-      directoryForDownloads: di.inject(directoryForDownloadsInjectable),
-    }),
-
-  lifecycle: lifecycleEnum.singleton,
-});
-
-export default installFromSelectFileDialogInjectable;
